PDF interpreter - allow unlimited growth of stacks
Bug #696487 "Error: /undefined in --run-- reading PDF file"

The problem is caused by the fact that we use a dictionary to store
ExtGstate changes, so every time we do a gsave we need a new dictionary
to hold those changes. The file is sub-optimal and contains thousands
of nested gsaves, which means we need thousands of dictionaries, and
quickly fill up the dictionary stack.

We could specify a larger maximum dictionary stack, but there is no
obvious limit to files of this type, one page was seen to have in excess
of 6,000 nested gsaves. Instead we permit unchecked stack growth by
setting MaxDictStack (or MaxOpStack / MaxExecStack) to -1 so that the
stack will grow as required until we run out of memory.

Bug #696511 has been opened as an enhancement to move the ExtGstate
parameters into the regular graphics state which has to be better than
the current kludge.

No differences expected.