PS/PDF interpreters - fix Text rendering mode 3 with large glyphs
Bug #696513 "Regression: Ghostscript renders unexpected boxes starting with bb246f03fab855325a73fdc982ed9802c1f16772"

The commit noted as causing the regression is, frankly, incorrect.

When handling text we have two possible paths; one with caching and,
for large glyphs at high resolution, one without caching.

When doing a stringwidth or Tr3 with a cached glyph, we simply avoid
rendering the cached bitmap. However with non-cached glyphs its not so
simple. We run the CharProc/BuildGlyph procedure as PostScript, which
means all the operations are indistinguishable from regular PostScript.

So, when performing a stringwidth using a non-cached glyph, and prior to
the noted commit, when using Tr 3, we push the null device in order to
prevent the glyphs actually marking the page, and then grestore back to
original device when we finish.

The commit noted broke that by failing to push the null device, meaning
that uncached glyph in rendering mode 3 did actually mark the raster.

The first part of the commit reverses the earlier commit, and implements
skipping the currentpoint by explicitly checking the 'operation' before
performing it. In fact this is no longer necessary as the PDF interpreter
now checks this at an earlier stage and the original bug can no longer
be reproduced this way anyway.

However, this exposed two new problems which had previously been masked,
but were nevertheless potential bugs:

The 'procs' structure in a device was originally intended to be fully
populated, each proc should either have a device-specific method or a
graphics library default. This means there is no need to check the value
of a device proc, we can always execute it.

Unfortunately, as noted before, this has been broken for some time in
the code base. There are several methods which do not have graphics
library defaults, and are not set up by the default device macros. This
can lead to the 'procs' members being NULL.

There are several places in the code where we explicitly check for a
device method being NULL before execution in order to catch this
condition. Ideally we should go back and fix these properly so that we
can go back to simply executing the device method without checking. But
this is a big job.....

So here we add two more cases where we check a device method
'update_spot_equivalent_colors' to be sure its not NULL before we
execute it.

No differences expected.