Bug 696466: Fix incorrect line joins in strokes.
When we flatten a path for stroking, we keep 'notes' on each
line to tell us whether each line was present in the original
path, or whether it was generated by the flattening process.

We further note which lines were generated by the flattening
process, and are NOT the first one in a given curve. We do this
so that we can apply the 'curve join' to the start of such line
segments.

The "curve_join" is an optimisation to speed path drawing. When
we've flattened a curve, it makes no sense to draw (say) a round
join between each flattened line, because they will be almost
parallel, and the join will be visually indistinguishable from
a simple bevel (in most cases).

It's still important that we draw the correct curve between any
preceeding line and the first line in the curve though.

Sometimes, (especially when setaccuratecurves is turned on), the
first line in a flattened curve may be degenerate (i.e. 0 length).
The stroker spots degenerate sections and skips them. Unfortunately
that means it also skips the note about it being the first line
from a curve, and the wrong join is used.

The fix implemented here is to alter the stroker so that when
it skips a degenerate section, it checks for the 'first line
from an arc' flag and carries it forwards.