pdfwrite - correct a buffer size when converting colours
Bug #696531 "Ghostscript segfaults on some PDF files"

When converting DeviceN (and some other) spaces to a different colour
space we do not convert the actual space, we convert the *base* space
in order to better preserve the original colour intent.

For example a Separation ink of 'PANTONE 4' might have a base space (to
which the PANTONE colour will be converted if that ink is not available)
of DeviceCMYK. If we set the ColourConversionStrategy to RGB then we can
either convert every pantone colour to CMYK then convert the CMYK to RGB
or we can alter the base space from CMYK to RGB, we choose to alter the
base space.

To do this we need to create a function which maps the input colours to
the base space. This function needs a data source, and the buffer being
allocated to hold the data had its size incorrectly calculated.

This led to running off the end of the buffer and corrupting memory.
This commit corrects the buffer size calculation.

No differences expected.