Fix badly scaled image (696530) and refactor.
Many RTL plotters do not restrict resolutions so we allow any resolution
in RTL mode.  Further the the function which sets the resolution was
awkward: wrongly commented with a complex nested ternary operator.  We
now use a simpler lookup table and hopefully have a better explanation
of what we are doing in the comments.