Bug 696508: improve configure cross compile support
This is a squashing of the commits on the branch:

CCAUX-configure

This allows CCAUX and related settings to be set on the configure command
line, or in the environment.

Also, using configure to guess at creating an arch.h file which gets used
rather than genarch generating one. A couple of things in that file will
not be as accurate as using genarch, but those are rare, and almost
certainly for outdated systems (such as whether floats use IEEE
representation).

Better documentation to follow, but as an example, building for Android would
work with:

./configure --host=arm-linux-gnu --build=x86_64-linux-gnu \
CC=/opt/android-ndk-r10e/toolchains/arm-linux-androideabi-4.9/prebuilt/linux-x86_64/bin/arm-linux-androideabi-gcc \
CFLAGS="--sysroot=/home/cliddell/bin/android-ndk-r10e/platforms/android-18/arch-arm -MMD -MP -MF -fpic \
-ffunction-sections -funwind-tables -fstack-protector -no-canonical-prefixes -march=armv7-a -mfloat-abi=softfp \
-mfpu=vfpv3-d16 -mthumb -fomit-frame-pointer -fno-strict-aliasing -finline-limit=64 -Wa,--noexecstack" LDFLAGS=" \
--sysroot=/home/cliddell/bin/android-ndk-r10e/platforms/android-18/arch-arm -MMD -MP -MF -fpic -ffunction-sections \
-funwind-tables -fstack-protector -no-canonical-prefixes -march=armv7-a -mfloat-abi=softfp -mfpu=vfpv3-d16 -mthumb \
-fomit-frame-pointer -fno-strict-aliasing -finline-limit=64 -Wa,--noexecstack" CCAUX=gcc CFLAGSAUX=" "