Get dictionaries working
Start by moving token parsing to its own routine so that we can call it
from both the top 'distillerparams' code and the dictionary code.

Now that dictionaries are (in outline at least) working, it looks like
it ought to be possible to massively simplify the way that arrays
work by defining them as a collection too.

Then we can do collections inside collections very easily, which we
need to do.