PDF interpreter - more robustness in the face of broken Info dicts
Bug #696541 "ps2pdf fails (typecheck) on a file that can be opened by some other viewers"

This only fails if the device states it can accept document info,
currently that's limited to pdfwrite and friends.

The problem is that the Info dictionary contains invalid strings, and
those cause the parsing to break with an error.

This commit simply ignores the error (thereby bit-bucketing the Info)
just like other devices.