Fix for Bug 696514. rinkj device
The rinkj device was dereferencing a NULL pointer related
to the ICC device link profile that it can be set up to use.
Fixing this issued and playing around with a various command
line options revealed a problem in error trapping in the
icc code where you could cause a crash by having a bogus
device link profile specified on the command line.  This
fixes that crash also.