Make waterfal.ps standalone
Originally waterfal.ps called .runlibfile to run landscap.ps (in the lib
directory) to draw the page landscape.

Since .runlibfile is non-standard Postscript, and lib is no longer in our
default search path, *and* it is only five lines of Postscript, it's a
needless complication - better for the file to be self contained.