PDF interpreter - Handle inline images in a text block
Bug #696545 "Missing inline images"

Similar to bug #695897, where image XObjects were being drawn inside
a text block (between BT and ET), in this case we have inline images
(ID) inside a text block.

This is, according to the specification, illegal but Acrobat simply
ignores that and draws the image using the CTM in force when the BT
operator is encountered.

This commit just reuses the code for Bug 695897 and applies it to the
ID operator as well.

No differences expected