Bug 696553: reduce redundant loading of (some) substitute fonts
Loading fonts from a Fontmap mapping, the actual font name from the font
file usually does not match the Postscript font name.

When that happens, a subsequent attempt to load the same font name will not
find the font in the font directory, and this triggers the whole font file
loading process via Fontmap.

Add code to create a new font with the original Postscript font name.

Also, undefine the orginal font from the font file, to avoid cluttering up
the FontDirectory with pointless entries.