pdfwrite - fix alpha duplication
Bug #696524 "gs: Some elements discarded in PDF with 3d transform"

The bug title is incorrect. In fact the problem is complex; we have two
forms, each of which is preceded by a graphics state, each graphics state
modifies the constant alpha, but one has alpha is shape false, while the
other has it as true.

This causes confusion when updating the alpha as we cannot have AIS
both true and false at the same time, so we don't know which alpha to
use.

This commit checks the graphics state to see which alpha usage is the
default state and sets AIS from that, as appropriate for whether the
opacity or shape is set.

This is a little hacky but I think its safe enough and fixes what is a
rather rare problem.