PS interpreter - fix array processing for the DCTEncode params dict
The PLRM says that HSamples, VSamples can be strings, arrays or packed
arrays and QuantTables can be an array or a packed array (of strings,
arrays or packed arrays). If they are array types, then the values must
be integers.

In sdcparam.c where we process the paramters we were checking and
processing strings, and if the data was not a string we tried to handle
it as an array of floats (which is technically incorrect), but we never
tried to process the data as an int array.

This led to the routine throwing a typecheck error, so the values were
never used.

This commit adds processing for int arrays (I've left the float array
processing intact, even though its wrong).

No differences expected, this is really only used by pdfwrite.