Bug 696516: tiffsep(1) validation of compression method
Use a meaningful value of bits-per-component to validate the compression setting
for tiffsep and tiffsep1.

Since both those devices ignore the device BitsPerComponent setting, we cannot
rely on that, so check explicitly for tiffsep or tiffsep1 and use an
appropriate bpc value depending which is in force.

Also, change the comment on the default tiffsep(1) device(s) BitsPerComponent
value from "Not used" to "Ignored" which is more accurate, since bpc can and
does change, but has no effect on those two devices.