PS interpreter - further fix for DCTEncode params arrays
commit 96f79a46a559af75995bf021cc85438c99509bbb missed the fact that,
when a parameter is returned from parameter processing we need to
return a positive code > 0, otherwise the caller assumes that there was
no error, but no key found either.

This actually makes the HSamples, VSamples and QuantTables work for
the DCTEncode filter.

This causes a *very* large number of diffs, because the pdfwrite (and
ps2write) devices now actually use the parameters defined for these values
resulting in small differences in DCT (JPEG) output.

Also the Quality Logic test 23-12E.ps tests this feature and now shows
a difference.