Revert and fix commit a7655b5d2fc42217eac71efd01f22fe3cca33d4a
In the previous commit I misunderstood the return value meanings from
s_DCT_byte_params leading to an incorrect default (and causing many
diffs)

This commit properly sets the return value so that, if the key is not
found we use the correct default.

Again, lots of diffs expected here.