VS2015: Workaround apparent C runtime bug.
The behaviour of _read appears to have changed in VS2015 in
a way that is contrary to the documentation. If a _read call
is given a single "return" keypress, it puts \n in the buffer
but returns 0 as the number of things read.

This, according to the documentation means 'EOF'. We therefore
have some sneaky code to guess when it's lying to us.