Bug 696132: Grid fit images within pattern fills.
To avoid white lines appearing between pattern fills, grid fit
images whenever they appear within such fills.

Gridfitting involves stretching any axis aligned image to
completely cover the pixels it touches.

The vast majority of this patch is to do with detecting that
we are in a pattern accumulation device in a nice way (allowing
for the possibilities of pdf14 accumulators etc).