PDF interpreter - move a device name check into a parameter check
Bug #696568 "Treatment of page labels"

Previously we were checking against the device being pdfwrite before
processing page labels, because these require special handling. This
stops it working with the DejaVu device (which I wasn't previously
aware of) so instead we use the current method of checking a device
parameter to see if the device wants to be given the Page Labels.

Because the pdfmark syntax apparently cannot handle the range of
possible schemes for PageLabels (see bug #692901) we don't handle
PageLabels as pdfmarks, but instead as a specific device parameter.

We won't change this unless someone can show conclusively that the
pdfmark syntax can handle all possible formats for PageLabels.

No differences expected