Update the documentation
Completely reworked the documentation for the high level devices, this
is now all contained in one HTML file.

Broke up the pdfwrite family so that we now deal with the various options
for PDF and PostScript output separately, and also deal with the various
parameters and methods of setting them for each language separately.

Removed the now obselete and unreferenced ps2ps2.htm and ps2pdf.htm