Extend copy_alpha to support 8 bit depth.
Currently copy_alpha (and copy_alpha_hl_color) can only accept
2 or 4 bits of alpha data. In order for the work on bug 696562
to proceed we need it to support 8 bits of alpha data. Extend
the 2 default implementations here.

Also update the docs to mention the new bit depths, and to mention
the existence of copy_alpha_hl_color at all.

Remove the line in the docs that claims copy_alpha will not be
called unless get_alpha_bits returns non-1, as this will shortly
not be true.