PDF interpreter - yet more robust error handlign with broken files
Bug #696540 "ps2pdf fails on a file that can be opened by some other viewers"

Two problems here.

First, the file has been truncated and garbage written, after the startxref
token. Previously we used the 'token' operator to try and read both the
'startxref' and the actual offset value. However, while executing the
token operator we reached EOF, and this *closes* the underlying file.
Unsurprisingly this then caused ioerrors on every subsequent operation.

So, define a new routine 'token_no_close' which installs a SubFileDecode
filter on top of the existing file/filter chain and reads from that. We
explicitly set CloseSource to false so that even if we encounter EOF
while executing the file we will only close the filter and not the
underlying file/filter chain.

However this then exposed a different problem when rebuilding the xref;
we scan backwards looking for a trailer with a /Root key. But if the
trailer was early in the file (< 64Kb), and the block reading worked
out so that the initial block was less than 64Kb then we would calculate
The offset to the trailer incorrectly because we assumed the block size
was always 64Kb, which it isn't if we don't have 64Kb to read.

No differences expected