Bug 696562: Interpolate imagemasks.
When rendering imagemasks, we can get nasty dropout effects if we
should be using interpolation. To fix this, we implement
interpolation for the non-high level device case.

We use copy_alpha (and copy_alpha_hl_color) to write the data. We
only do this in the case of devices that don't declare themselves
to be high-level.