silence a couple of compiler warnings
Check and action the return value from fread throw an ioerror if it
failed.

Prototype a couple of functions (this is an old warning)

I did  not remove two functions 'pjl_impl_set_envvar' and
'pjl_impl_set_defvar'. Although these are not used at present they
match 'pjl_impl_get_envvar', which *is* used. It would be surprising
to a programmer if only one accessor was defiend.