Detect 'impossible' code path being called.
As part of the work towards bug 696562, we spotted a code path
that can never be run.

gx_alloc_char_bits is only called from one place (base/gxchar.c
line 607). dev2 can only be non-NULL if:

  iwidth > MAX_CCACHE_TEMP_BITMAP_BITS / iheight &&
  log2_scale.x + log2_scale.y > alpha_bits

The second half of that condition can never be true due to the
construction of the log2_scale.{x,y} values in
gx_compute_text_oversampling().

Rather than rip this code out just before a release (and inevitably
be proved wrong), we add an error in this code path. If no one
reports it to us, we'll rip the code out later.