PDF interpreter - handle Xref stream with incorrect length
Bug 696560 "Error reading PDF file"

The file uses compressed Xref streams (several of them in fact) and one
of them has an incorrect length (36 instead of 34). This leads to the
interpreter reading past the end of the stream and consuming the first
two bytes of the following token, so instead of seeing 'endstream' as
we should, we encounter 'dstream'.

We already have code which attempts to detect and handle this condition
but it assumes that the stream is being executed as an object, not as
an xref, and expects the object number and generation to be on the
stack, which they are not when we are dealing with an Xref stream.

In this commit, when we detect such a problem we attempt to verify if
we are dealing with an object by checking to see if the stack contains
at least 3 objects (the stream dict, the object and generation) and that
the second and third objects down are integers. If these conditions are
met we end the object and execute the 'exit' operator to terminate
the calling .pdfrun function. Otherwise we simply return with the
dictionary on the stack, as this is what the xref code expects.

The regression file Bug696240.pdf should now render correctly.