Bug 696132/696572: Continued "pattern gap" work - interpolate
Marcos reports (in bug 696572) that many files have problems
after the commit done for bug 696132. That commit was intended to
cause all orthogonal images within a pattern to be gridfitted.
Due to a mistake, this actually caused all orthogonal images
ANYWHERE to be gridfitted. That bug was fixed a few commits
ago (c902d4d).

Having fixed that, so we now operate as originally intended
we still have some problems showing. These (in the cases tested
at least) appear to be due to the radically different renderings
given of images when they are very slightly permuted. This is
because when images are being rendered at a smaller size
than their source, dropouts occur.

A classic example of this might be an image of the form:

  * * * * * *
   * * * * *
  * * * * * *
   * * * * *

which is used as a pattern to give a halftone efferct.

When downscaled, this is very sensitive to dropouts; we can easily
get solid black or white out of such images. Using interpolation
gives a truer rendition of the intended output.

This commit spots such downscales of gridfitted images within
pattern accumulators and forces interpolation on for them.