PDF interpreter - allow the operand stack to gro without limit
Bug #696567 "Error reading PDF file: /stackunderflow in --run--"

The file has a page with an outrageous number of ExtGStates on it,
447000, many (possibly all) of which do not contain any actual gstate.

This was causing a stackoverflow because the maximummoperand stack
size was set at 300,000. This commit builds on the code for
Bug #696487 (from the same customer) and enables the operand stack to
grow without limit when interpreting PDF files.

Of course, the page is still stupidly slow.

No differences expected.