pdfwrite - disallow missing glyphs in CIDFonts when creating PDF/A
Bug #696547 "Converting to PDF/A using -dPDFACompatibilityPolicy=2 returns without error but produces invalid PDF/A files"

One part of this bug. PDF/A insists that all fonts be embedded, and
all glyphs used from that font are present. If we detect a condition
where that is not the case, and we are emitting PDF/A consult the
CompatibilityPolicy and take appropriate action.