Bug 696132: Reapply previous work.
Previous commits in this area went in piecemeal and caused lots of
diffs which were then fixed etc. By committing here in one go, we
hope to get saner results from the tests which we can have more
confidence in.

A summary of the changes:

1) Introduce a mechanism so that we can know whether we are in
a pattern accumulator or not.

2) When plotting an othogonal image use that mechanism to detect
whether we are in a pattern accumulator. If we are, then grid
fit the image.

3) If we are in a pattern accumulator and we are downscaling an
image, then interpolate it. This avoids nasty dropouts with
halftoned images that can (due to nearest neighbour plotting)
suddenly change massively in appearance.