graphics library - make vector devices implement the correct spec_op handler
Devices based on gdevvec were not properly having the spec_op method
assigned to the gdevvec spec_op handler if the device didn't handle
spec_op itself.

Ideally this would be done by having the device prototype being fully
populated, but the number of device methods makes that impractical.
Instead, if the method is NULL or the gx_default method, replace it
with the gdevvec one.

This already works for devices based on gdevprn as a simlar hack is
performed in gx_default_create_buf_device

No differences expected