Bug 696323: Fix indeterminism in thresholding code.
The thresholding code calculates the destination width of the
image lines in one way, then fills the data based on the
bresenham. It is possible for the destination width calculated
to not match that given by the bresenham, in which case
undefined data can be plotted.

The fix here is simply to calculate the destination width from
the bresenham, so it will exactly match the actual data we get.