pdfwrite - force non-Identity CMap emission when creating PDF/A
Bug #696547 "Converting to PDF/A using -dPDFACompatibilityPolicy=2 returns without error but produces invalid PDF/A files"

The third and (so far) final bug in this collection.

PDF/A mandates that all CMaps except Identity-H and Identity-V must be
embedded in the PDF file. Previously we were not doing that for the
'standard' CMaps.

This commit forces the emission of all CMaps except Identity ones when
producing PDF/A output.

No differences expected