Further indeterminism fixes for halftoning.
Yesterday I changed the halftone code to calculate the dest_width
etc from the bresenham. Now it turns out that the bresenham is
tweaked slightly in some cases, so that the 2 places where it
was being used to calculate dest_width were still getting
different results.

This commit rearranges the code so that the same bresenham values
are used consistently, and so should give us proper matches.