pdfwrite - don't render shadings if we don't need to convert base space
Bug #690125 "Gradient / Fill Pattern Conversion Issues"

Since the change to ICC colour management, the check to see whether we
could support the base space of a shading was incorrect, it checked the
strategy against 'ICC' instead of the real base space of the pattern.

This change checks the actual base space of the pattern before deciding
whether we can handle it.

Although this 'fixes' this bug its not the whole story, I want to alter
the functions so that they can be sampled and generate a different colour
space if we are doing colour conversion, rather than rendering the
shading.

No differences expected.