pdfwrite - fix a colour conversion crash
Discovered while working on an unrelated problem. If we had set the
CompatibilityLevel to < 1.3 and encountered a CIEBased colour space in
the input file, then we unconditionally return a 'convert' result
because we cannot embed an ICC profile in such an old PDF version.

However we did not update the 'pcs_orig' argument, and later tried to
use it, resulting in a NULL dereference.

No differences expected.