Bug 696609: Fix x11alpha regressions.
The changes to pattern_accum_get_bits_rectangle were causing
a lack of output.

This was due to the call to the underlying get_bits_rectangle
returning with a pointer to the data (rather than supplying
a copy of it). The blank_unmasked_bits call would then fail
as it could not alter the underlying data safely.

The fix is simply to remove the bit that allows a copy of the
data to be submitted from the options before calling, thus
ensuring that we can operate on a copy.