Bug696540: Fix pattern accumulator initialisation
When we create a pattern accumulator the bitmap contents are
initially undefined. We then draw a rectangle over them to set
them to known values.

Unfortunately the code that writes this bitmap does not check
for the ctm being sane, so in some cases the initialisation
can fail.

This shows up as indeterminisms in the alpha blending.

The simple fix is to set the ctm to the identity matrix before
rendering.