Remove unused variables/conditions from gdevpdf.c
Spotted while debugging bug 696612.

Various places in the code do:

{ int j, code = 0;
  for (j = 0; j < n && code >= 0; j++)
  {
    STUFF
  }
}

which is a perfectly reasonable thing to do - except for the
facts that: 1) STUFF never alters the value of code, and 2) even
if STUFF did alter the value of code, we never check the value
and return it.

Accordingly, I've just removed the references to code.