Bug 696612: Swallow rangecheck errors in psf_check_outline_glyphs
At the closedown of the file, we run through and write out fonts.
As part of this process, we check the glyphs in the font. If any
of the glyphs come back as bad, we abort the whole process.

Previously we ignored any errors here, and my change to make us
not ignore errors in the pdf_close routine caused this regression.

After discussion with Ken and Chris, the correct fix, I believe,
is to continue to catch and honour all errors in pdf_close, but
to explicitly swallow certain errors lower down.

Chris suggested, and I agree with him, that simply swallowing
the rangecheck error in psf_check_outline_glyphs would be an
acceptable fix (for now at least).

I am leaving the bug open and passing it to Ken so that he can
double check this area in more detail at his convenience.