Bug 696603: Quieten address sanitiser.
mem_mono_copy_mono includes some clever code that reads
16 bits at a time and shifts to do fast mono copies of
unaligned data. This can result in overreading the end
of data by a byte, but never so far as to cause address
overflows due to the granularity at which data can be
allocated.

The 'overread' data is never actually used.

The simple fix here just extends the source block by
a byte to avoid error sanitizer complaining. Valgrind
correctly does not flag this.