Bug 696611: Avoid imagemask interpolation errors with pbmraw.
pbmraw (deliberately) doesn't know how to copy_alpha. We never
call copy_alpha when going to pbmraw direct because of this, but
the clist wrapping defeats our detection. This results in an
error.

Tests enabling copy_alpha for pbmraw show a degredation in render
quality, so instead, we just disable all imagemask interpolation
when going to halftone devices.