high level forms - account for flipped/mirrored CTM
For high level forms support, if the device requests a specific CTM be
applied we calculate a clip path which includes negative co-ordinates
to ensure that a translated form won't be erroneously clipped to the
page.

When doing this, we need to account for the CTM when the form is executed
potentially being flipped or mirrored (or both!)

No differences expected.