Bug 696616: Move antidropout downscaler functionality.
The current code globally enables use of the antidropout
downscaler on all halftoned devices (unless they override
the gxdso call), when interpolation is enabled.

We now introduce a -dAntidropoutDownscaler option that
defaults off. That sets a bit in dev.color_info that is
used to control the option. This decouples it from
interpolation.