pdfwrite fix address sanitizer complaint.
When checking a font name to see if its a URW replacement for a base 14
font we were doing a memcmp with the length of the candidate name.

If that length exceeded the length of the URW name then we were in
effect comparng against part of the next name (or random bytes off the
end of the table.

Of course this is harmless except in the highly unlikely case of the end
of the table not being followed by more data bytes. But the address
sanitizer complains. So we now compare the length of the two strings
first.