ps2write - suppress an address sanitizer warning for MM fonts
ps2write does not currently handle Multiple Master fonts well, we know
this there is an open enhancement for it. We do have a hacky work-around
which turns the MM OtherSubrs into regular OtherSubrs using the blended
values.

This doesn't work when the font program uses constrcuts such as ' x y div'
in order to create floating point values. This could mean that we did
not have as many paramters as expected on the stack as are required for
a given MM OtherSubr, leading to us indexing off the bottom of the
stack.

 This meant we were using random values, but this didn't really matter
as the data is always going to be wrong. However the address sanitizer
complains about this.....

In this commit if we would underrun the stack we just write a 0 instead