pdfwrite - FunctionType 2 C0 and C1 are optional
Bug #696626 " A PDF file causes ps2pdf crash"

The code fro serialising a Type 2 function was assuming that C0 and C1
were valid (validated by the interpreter) and always present. In fact
both of these are optional. We were attempting to dereference a NULL
pointer.

Altered the type 2 serialise code to write the default values if either
or both of C0 and C1 are not present.

No differences expected.