Fix psdcmyk handling of output to /dev/null or equivalent.
The PSD file format does not support multiple images (pages) in a single file,
so we spot such an attempt (by checking whether or not the output file string
has a "%d" in it) and throw an error.

This also throws an error if an attempt is made to write multiple pages to
/dev/null (as is done for performance testing).

Since it really doesn't matter if the output "file" is invalid when we're just
discarding the data, spot this case, and allow multi-page files to run
without error.