pdfwrite - fix writing of composite objects
When emitting composite objects we attempt (for reasons which are
unclear) to do so in chunks of 256 bytes or less. The code to split a
large buffer into smaller buffers was simply broken, did not work
properly at all.

This commit fixes the logical flaw, allowing us to write composite
objects larger then 254 bytes without error.

NB this is a rare condition, such objects are usually only created
from input such as pdfmarks, normal composite objects are created and
emitted by pdfwrite itself and do not go through this code.