pdfwrite - more changes for composite object emission
This 'fixes' Bug 696657, but it is not correct in general.

We break the emission of composite objects up into blocks of less than
255 characters for compliance with DSC output, but if any one of the
objects contains certain characters ('/' '[' '{' '(' or ' ') then we
might break the object inappropriately (this is likely for strings and
is the case for ths file).

There is no way to address this 'properly' in general, its perfectly
possible to create a PostScript object more than 255 bytes which
cannot be broken up, so I'm going to ignore the problem for now, since
it only occurs when writing pdfmarks to the ps2write device and having
DSC turned on (which is the default).

If anyone really needs to do this, they should set ProduceDSC to false.