pdfwrite - optimise rectangular subpaths as 're' where possible
The current 'vector device' path code doesn't work properly with clips and
doesn't optimise rectangular subpaths as rectangles. (For the purposes
of 're' optimisation a rectangular subpath is a moveto, followed by 4
lineto's at right angles, followed by a closepath).

This commit creates new code to handle pahts in pdfwrite, if the enumerator
is a clip enumerator it correctly enumerates the subpaths from the clip
instead of enumerating the path.

We buffer up operations looking for rectangles; when we find one, if the
current matrix is not sheared or skewed, then we check to see if the
rectangle is 'x first'. If it isn't we move the starting point round by one
as 're' always emits width first. We also check to see if we are doing
a dashed stroke, as if we are we cannot move the initial vertex as this
will cause the dash to be incorrect (abort the optimisation).

We copy the code from gdevvec whcih checks to see if the whole path is
a rectangle, and also the 'optimise' code whch will concatenate colinear
moveto and lineto operations into a single operation.

Finally we also preserve the heuristic that prevents a fill with a trailing
moveto as ths apparently exposed a bug in old versions of Acrobat.

Ths does result in a number of single pixel or single pixel wide
differences in the test suite, but they all look like simple differences
to me, not regressions. The new code doesn not appear to be
singificantly slower and ths does produce a useful reduction in file
size for some files.