PDF interpreter + pdfwrite - use font object numbers to detect identical fonts
Previously we have had problems when dealing with multiple input PDF files
where the files contain subset fonts with poorly chosen subset prefixes.
Because we only have the font name to work with it was difficult to
determine whether two fonts with the same name were, in fact, the same font.
Incorrectly deciding that fonts were the same led to glyphs being wrong
in the pdfwrite output,  whereas treating all fonts as unique resulted in
outputting far, far too many font instances.

In ths commit we add a means to determine from PostScript the filename
associated with a PostScript file object and we use that to get the
filename of the current PDF file being processed in the PDF interpreter.

When we encounter a font in a PDF file, we check to see whether it has
an object number and a FontDescriptor, if the FontDescriptor is present
we get its object number. We then check to see if the FontDescriptor
has a FontFile, if it does not then the font is not embedded.

If the font is embedded then we use the object number of the FontDescriptor
(two fonts can have the same FOntDescriptor, and we can treat these as
being the same font) and combine it with the PDF filename using a
CRC to produce a hash which we store in the font as an XUID.

pdfwrite then uses the XUID (if present) to determine if two fonts
are in fact the same font.

Adding the XUID causes a very small number of pixel differences in
the text of some PDF files (especially, for some reason, with the
psdcmy device). Also, a few PDF files have multiple copies of the
same font, with different FontDescriptors, which were previously merged
but now are not, whch leads to a small increase in file size.

This new behaviour can be disabled by adding the -dPDFDontUseFontObjectNum
switch on the command line.