txtwrite - reinstate an earlier commit by Robin Watts
This was removed as I couldn't, at the time, be certain that it was not
required. Having spent some time working through the various macros, and
the code, I'm now convinced that this is not required, and is probably
a remnant of an early attempt at this device, which used garbage collected
objects.

This was originally added by Robin as a small part of commit
fd9a66f997bb57e9628a703774eddcf933475a34 and removed by me in commit
15b3f5cbf12461e2ed318e793669b7c34e32089b


No differences expected