PDF Interpreter - try to work around invalid ICCBased spaces
Bug #696690 and Bug #696120

These files have ICCBased spaces where the value of /N and the number of
components in the profile differ. In addition the profile in Bug #696690
is invalid. In the case of Bug #696690 the /N value is correct, and the
profile is wrong, in the case of Bug #696120 the /N value is incorrect
and the profile is correct.

We 'suspect' that Acrobat uses the fact that Bug #696120 is a pure image
to detect that the /N is incorrect, we can't be sure whether it uses the
profile or just uses the /N to decide on a device space.

What we now do is; If the /N and device profile number of components
don't match, we assume the device profile is correct and patch /N to be
the same as the profile (see /ICCBased-resolve), but we save the
original value of /N in /OrigN. In the PostScript setcolor operator, if
the space is a genuine ICCBased space (not a replacement for a device
space) we call set_dev_color which will actually exercise the profile.
If that fails we return an error. For ICCbased spaces we run setcolor in
a stopped context, and if it fails we check to see if there is a /OrigN
(this occurs only if the /N was different to the number of components in
the profile). If there is a /OrigN then prefer that to the profile,
otherwise they agreed, so just use /N and select a device space. If we
can't select a device space with the correct number of components, give
up and throw an error.

This produces some differences in a small number of files, mostly the
PAM device at 72 dpi. These are not visually perceptible, and since they
only occur on one device at low resolution I think they are spurious.